Implimented pagination
In the listing api function added currentpage to fetch data based on the page also updated the litmit from 20 to 15

Moved the fetching query form the index.tsx page to the listings page. This way i don't have to drill the current page state 2 layes below. Also i can just pass the category prop form the index to the listings

In the pagination component added onclicks to handle previous and next page as well as disbling the buttons when certain conditions are met, to prevent users from accidentaly putting a negative page or more than available page